refactor(combo): revert Ivy workaround in combo click handler
