+++ projects/igniteui-angular/src/lib/combo/combo-add-item.component.ts
@@ -1,5 +1,5 @@
 import { IgxComboItemComponent } from './combo-item.component';
-import { Component, HostListener } from '@angular/core';
+import { Component } from '@angular/core';
 
 /**
  * @hidden
@@ -17,18 +17,10 @@ export class IgxComboAddItemComponent extends IgxComboItemComponent {
     }
 
     /**
-     * @hidden
-     * @internal
-     * This is related to https://github.com/angular/angular/issues/33300
-     * When the above is fixed, we can remove the @HostListener decorator and move
-     * the body of the `handleClick` method back under `clicked`
+     * @inheritdoc
      */
-    @HostListener('click')
-    handleClick() {
+    clicked(event?) {
         this.comboAPI.disableTransitions = false;
         this.comboAPI.add_custom_item();
-    }
-
-    clicked(event?) {
     }
 }
